rename file to lastUsedPaths
