+++ src/renderer/components/dialogs/CreateChat.tsx
@@ -54,7 +54,7 @@ import useConfirmationDialog from '../../hooks/useConfirmationDialog'
 import {
   LastUsedSlot,
   rememberLastUsedPath,
-} from '../../utils/cachedLastUsedPath'
+} from '../../utils/lastUsedPaths'
 import { dirname } from 'path'
 
 type ViewMode = 'main' | 'createGroup' | 'createBroadcastList'
