+++ src/renderer/components/dialogs/EditProfileDialog/ProfileImageSelector.tsx
@@ -6,7 +6,7 @@ import useTranslationFunction from '../../../hooks/useTranslationFunction'
 import {
   LastUsedSlot,
   rememberLastUsedPath,
-} from '../../../utils/cachedLastUsedPath'
+} from '../../../utils/lastUsedPaths'
 import { dirname } from 'path'
 
 export default function ProfileImageSelector({
